requirements.txt
Django==4.2.6
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.100.0
pytest==7.4.2
pytest-django==4.5.2
requests-mock==1.11.0
python-dotenv==1.0.0
